Add dropped metrics in the process part of the agent
